Move collector functions to seperate files
