+++ repository/collectors/meeting-collection.js
@@ -1,0 +1,80 @@
+import { updateTriplestore } from '../triplestore';
+
+/**
+ * Helpers to collect data about:
+ * - meetings
+ * - newsletters
+ */
+
+/*
+ * Collect related meetings for the relevant agenda's
+ * from the distributor's source graph in the temp graph.
+ */
+async function collectMeetings(distributor) {
+  const properties = [
+    [ 'besluitvorming:isAgendaVoor' ], // meeting
+    [ '^besluitvorming:behandelt' ], // meeting
+  ];
+  const path = properties.map(prop => prop.join(' / ')).map(path => `( ${path} )`).join(' | ');
+
+  const relatedQuery = `
+      PREFIX besluitvorming: <http://data.vlaanderen.be/ns/besluitvorming#>
+      PREFIX ext: <http://mu.semte.ch/vocabularies/ext/>
+      INSERT {
+        GRAPH <${distributor.tempGraph}> {
+          ?s a ?type ;
+             ext:tracesLineageTo ?agenda .
+        }
+      } WHERE {
+        GRAPH <${distributor.tempGraph}> {
+          ?agenda a besluitvorming:Agenda ;
+              ext:tracesLineageTo ?agenda .
+        }
+        GRAPH <${distributor.sourceGraph}> {
+          ?agenda ${path} ?s .
+          ?s a ?type .
+        }
+      }`;
+  await updateTriplestore(relatedQuery);
+}
+
+/*
+ * Collect related newsletters for the relevant meetings
+ * from the distributor's source graph in the temp graph.
+ *
+ * Newsletters are only copied if they have already been published
+ * I.e. triple ?meeting ext:heeftMailCampagnes / ext:isVerstuurdOp ?sentMailDate . exists
+ */
+async function collectReleasedNewsletter(distributor) {
+  const properties = [
+    [ 'ext:algemeneNieuwsbrief' ], // newsletter
+  ];
+  const path = properties.map(prop => prop.join(' / ')).map(path => `( ${path} )`).join(' | ');
+
+  const relatedQuery = `
+      PREFIX prov: <http://www.w3.org/ns/prov#>
+      PREFIX besluit: <http://data.vlaanderen.be/ns/besluit#>
+      PREFIX ext: <http://mu.semte.ch/vocabularies/ext/>
+      INSERT {
+        GRAPH <${distributor.tempGraph}> {
+          ?s a ?type ;
+             ext:tracesLineageTo ?agenda .
+        }
+      } WHERE {
+        GRAPH <${distributor.tempGraph}> {
+          ?meeting a besluit:Vergaderactiviteit ;
+              ext:tracesLineageTo ?agenda .
+        }
+        GRAPH <${distributor.sourceGraph}> {
+          ?meeting ext:heeftMailCampagnes / ext:isVerstuurdOp ?sentMailDate .
+          ?meeting ${path} ?s .
+          ?s a ?type .
+        }
+      }`;
+  await updateTriplestore(relatedQuery);
+}
+
+export {
+  collectMeetings,
+  collectReleasedNewsletter
+}
